Feature: Alter scheduleEvent

+++ app/src/main/java/com/lifeisle/jekton/bean/ScheduleEvent.java
@@ -32,7 +32,7 @@ public class ScheduleEvent implements Cloneable {
 
     private static final String TAG = ScheduleEvent.class.getSimpleName();
 
-    public int id;
+    public long id;
     public String title;
     public long startMillis;
     public long endMillis;
@@ -84,7 +84,7 @@ public class ScheduleEvent implements Cloneable {
 
     public static ScheduleEvent newInstance(JSONObject jsonObject) throws JSONException {
         ScheduleEvent event = new ScheduleEvent();
-        event.id = jsonObject.getInt(ScheduleContract.OWN_EVENT_LOCAL_ID);
+        event.id = jsonObject.getLong(ScheduleContract.OWN_EVENT_LOCAL_ID);
         event.title = jsonObject.getString(ScheduleContract.OWN_EVENT_TITLE);
         event.startMillis = jsonObject.getLong(ScheduleContract.OWN_EVENT_START_TIME);
         event.endMillis = jsonObject.getLong(ScheduleContract.OWN_EVENT_END_TIME);
@@ -98,7 +98,7 @@ public class ScheduleEvent implements Cloneable {
 
     public static ScheduleEvent newInstance(Cursor cursor) {
         ScheduleEvent event = new ScheduleEvent();
-        event.id = cursor.getInt(ScheduleDBUtils.COL_EVENT_ID);
+        event.id = cursor.getLong(ScheduleDBUtils.COL_EVENT_ID);
         event.title = cursor.getString(ScheduleDBUtils.COL_EVENT_TITLE);
         event.startMillis = cursor.getLong(ScheduleDBUtils.COL_EVENT_START_TIME);
         event.endMillis = cursor.getLong(ScheduleDBUtils.COL_EVENT_END_TIME);
